feat(components): add defineComponent api
