+++ src/components/index.ts
@@ -1,0 +1,10 @@
+import defineComponent from './defineComponent';
+
+export * from './actions';
+export * from './config';
+export * from './hooks';
+export * from './properties';
+export * from './states';
+export * from './slots';
+
+export { defineComponent }
