+++ src/components/slots.ts
@@ -1,0 +1,11 @@
+
+export type SlotTemplate = {
+  component?: string;
+  slot: string;
+  key?: string;
+  props?: any;
+  code?: string;
+  children?: SlotTemplate[];
+};
+
+export type DynamicSlotTemplate = SlotTemplate | ((props: any) => SlotTemplate);
