--- src/components/state.ts
@@ -1,1 +1,0 @@
-export type StatePatcher<S = any> = (oldState: S) => S | S;
