+++ src/components/states.ts
@@ -1,0 +1,7 @@
+export type StatePatcher<S = any> = (oldState: S) => S | S;
+
+
+export type ComponentStateDefinition = {
+  label: string;
+  key: string;
+};
