+++ src/service/PageService.ts
@@ -1,4 +1,4 @@
-import { StatePatcher } from '../components/state';
+import { StatePatcher } from '../components/states';
 import IComponentService from './ComponentService';
 
 declare namespace IPageService {
@@ -8,6 +8,10 @@ declare namespace IPageService {
 declare class IPageService {
   static getService(pageCode: string): IPageService;
 
+  /**
+   * 根据传入的组件ID，返回页面中指定组件的前端组件服务。
+   * @param compId 组件ID
+   */
   getComponentService(compId: string): IComponentService;
 
   getComponentState<T = any>(compId: string): T;
