+++ src/types/entity.ts
@@ -4,10 +4,15 @@ export type EntityHttpResponse<T = any> = {
   data: T;
   error?: ResponseError;
 }
-
+/**
+ * EQ(相等) / NOT_EQ(不相等)：支持number类型；
+ * CONTAINS(包含) / NOT_CONTAINS(不包含)：支持enum类型包含条件、string类型模糊匹配
+ * GT(大于)/GTE(大于等于)：支持number和日期类型
+ * LT(小于)/LTE(小于等于)：支持number和日期类型
+ */
 export type RecordFilterItem = {
   fieldApiKey: string;
-  operator: string;
+  operator: 'EQ' | 'NOT_EQ' | 'CONTAINS' | 'NOT_CONTAINS' | 'GT' | 'GTE' | 'LT' | 'LTE';
   value: any;
 }
 
@@ -32,7 +37,6 @@ export type GetRecordByFilterRequestParams = {
       count?: boolean;
     };
   };
-
 }
 
 export type RecordData = {
